Add pull-up resistor for the soft-start circuit
The microcontroller can never bring the voltage on the gate of the soft-start
MOSFET to the 5V needed to turn off power. To only way to achieve 5V is to have
a pull-up resistor, and let the control pin in high-impedance mode.
When the control pin is configured as an output, and (preferrably) set low, the
osciboost module will be powered on. To power off the module, the pin is put
into high-impedance mode.

Add a 10k pull-up resistor to enable the power-off circuit to function
correctly.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>