Move BNC shadow from silkscreen to drawings layer
The swadow of the BNC plug in the molex footprint was placed on the silkscreen
layer. This caused the shadow, which is outside the PCB edges, to show up on
the silkscreen when generating gerbers. This could potentially confuse PCB
manufacturers.

Move the shadow of the BNC connector from the silkscreen layer to the drawings
layer.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>