Replace OPAMP with OPA1662 in MSOP-8 package
The previously selected opamps could only handle 6V, not the 10V we are feeding
them. Change the opamps to OPA1662, which have the desired characteristics.
Also, using dual opamps allows us to reduce the number of filter capacitors.

Because the change in footprint, re-laying part of the board was also needed.
The advantage is that the pinout of the MSOP-8 package is identical with many
other opamps.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>